+++ apps/desktop/src/window.ts
@@ -55,7 +55,6 @@ export function createWindow({serverUrl, rendererUrl, iconsDir}: CreateWindowOpt
       sandbox: false,
       contextIsolation: true,
       nodeIntegration: false,
-      webviewTag: true,
     },
   });